+++ packages/ui/FloatingLayer/FloatingLayer.js
@@ -10,10 +10,14 @@ import Cancelable from '../Cancelable';
 import {contextTypes} from './FloatingLayerDecorator';
 import Scrim from './Scrim';
 
-const forwardDismiss = adaptEvent(
-	() => ({type: 'onDismiss'}),
-	forward('onDismiss')
+const forwardWithType = type => adaptEvent(
+	() => ({type}),
+	forward(type)
 );
+
+const forwardDismiss = forwardWithType('onDismiss');
+const forwardClose = forwardWithType('onClose');
+const forwardOpen = forwardWithType('onOpen');
 
 /**
  * {@link ui/FloatingLayer.FloatingLayerBase} is a component that creates an entry point to the new
@@ -113,12 +117,14 @@ class FloatingLayerBase extends React.Component {
 	constructor (props) {
 		super(props);
 		this.node = null;
+		this.state = {
+			nodeRendered: false
+		};
 	}
 
 	componentDidMount () {
-		const {open, onOpen} = this.props;
-		if (open && onOpen) {
-			onOpen({});
+		if (this.props.open) {
+			this.renderNode();
 		}
 
 		if (this.context.registerFloatingLayer) {
@@ -126,13 +132,19 @@ class FloatingLayerBase extends React.Component {
 		}
 	}
 
-	componentDidUpdate (prevProps) {
-		const {open, onClose, onOpen, scrimType} = this.props;
+	componentWillReceiveProps (nextProps) {
+		if (!this.props.open && nextProps.open && !this.state.nodeRendered) {
+			this.renderNode();
+		}
+	}
 
-		if (prevProps.open && !open && onClose) {
-			onClose({});
-		} else if (!prevProps.open && open && onOpen) {
-			onOpen({});
+	componentDidUpdate (prevProps, prevState) {
+		const {open, scrimType} = this.props;
+
+		if (prevProps.open && !open) {
+			forwardClose(null, this.props);
+		} else if (!prevProps.open && open || (open && !prevState.nodeRendered && this.state.nodeRendered)) {
+			forwardOpen(null, this.props);
 		}
 
 		if (scrimType === 'none') {
@@ -145,7 +157,12 @@ class FloatingLayerBase extends React.Component {
 	}
 
 	componentWillUnmount () {
-		this.node = null;
+		if (this.node) {
+			const floatingLayer = this.context.getFloatingLayer();
+			floatingLayer.removeChild(this.node);
+			this.node = null;
+		}
+
 		off('click', this.handleClick);
 
 		if (this.context.unregisterFloatingLayer) {
@@ -179,31 +196,33 @@ class FloatingLayerBase extends React.Component {
 	}
 
 	renderNode () {
-		const {floatLayerClassName, open} = this.props;
+		const {floatLayerClassName} = this.props;
 		const floatingLayer = this.context.getFloatingLayer();
 
-		if (!this.node && floatingLayer && open && typeof document !== 'undefined') {
-			invariant(
-				this.context.getFloatingLayer,
-				'FloatingLayer cannot be used outside the subtree of a FloatingLayerDecorator'
-			);
+		if (this.node || !floatingLayer) return;
 
-			this.node = document.createElement('div');
-			floatingLayer.appendChild(this.node);
-			on('scroll', this.handleScroll, this.node);
-		}
+		invariant(
+			this.context.getFloatingLayer,
+			'FloatingLayer cannot be used outside the subtree of a FloatingLayerDecorator'
+		);
 
-		if (this.node) {
-			this.node.className = floatLayerClassName;
-			this.node.style.zIndex = 100;
-		}
+		this.node = document.createElement('div');
+		this.node.className = floatLayerClassName;
+		this.node.style.zIndex = 100;
 
-		return this.node;
+		floatingLayer.appendChild(this.node);
+		on('scroll', this.handleScroll, this.node);
+
+		// render children when this.node is inserted in the DOM tree.
+		this.setState({nodeRendered: true});
 	}
 
 	render () {
-		const props = Object.assign({}, this.props);
-		const {children, open, scrimType, ...rest} = props;
+		const {children, open, scrimType, ...rest} = this.props;
+
+		if (!open || !this.state.nodeRendered) {
+			return null;
+		}
 
 		delete rest.floatLayerClassName;
 		delete rest.floatLayerId;
@@ -212,18 +231,12 @@ class FloatingLayerBase extends React.Component {
 		delete rest.onDismiss;
 		delete rest.onOpen;
 
-		const node = this.renderNode();
-
-		return (
-			open && node ?
-				ReactDOM.createPortal(
-					<div {...rest}>
-						{scrimType !== 'none' ? <Scrim type={scrimType} onClick={this.handleClick} /> : null}
-						{React.cloneElement(children, {onClick: this.stopPropagation})}
-					</div>,
-					node
-				) :
-				null
+		return ReactDOM.createPortal(
+			<div {...rest}>
+				{scrimType !== 'none' ? <Scrim type={scrimType} onClick={this.handleClick} /> : null}
+				{React.cloneElement(children, {onClick: this.stopPropagation})}
+			</div>,
+			this.node
 		);
 	}
 }