+++ packages/ui/CHANGELOG.md
@@ -2,11 +2,31 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `ui/MarqueeDecorator` to stop marqueeing when using hover and pointer hides
+
+### Added
+
+- `ui/FloatingLayer.FloatingLayerBase` export
+
+### Changed
+
+- `ui/FloatingLayer` to call `onOpen` only after it is rendered
+
+## [2.0.0-beta.6] - 2018-06-04
+
+### Fixed
+
+- `ui/FloatingLayer` to render correctly if already opened at mounting time
+
 ## [2.0.0-beta.5] - 2018-05-29
 
 ### Added
 
-- `ui/FloatingLayerDecorator` imperative API to close all floating layers registered in the same id 
+- `ui/FloatingLayerDecorator` imperative API to close all floating layers registered in the same id
 - `ui/ProgressBar` and `ui/Slider` prop `progressAnchor` to configure from where in the progress bar or slider progress should begin
 - `ui/Slider` prop `progressBarComponent` to support customization of progress bar within a slider
 - `ui/ForwardRef` HOC to adapt `React.forwardRef` to HOC chains